# docs/api/game.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Game API",
    "description": "Bracket submission, match-result settlement, leaderboards, syndicates, Verified-Pundits.",
    "version": "0.1.0",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "paths": {
...
    "/v1/match/{match_id}/result": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "match_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3360",
      "description": "local dev"
    },
    {
      "url": "https://vtorn-game.aiva.nz",
      "description": "dev tunnel"
    }
  ],
...
}